requirements.txt
python-fasthtml
litellm
numpy
plotly
matplotlib